requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.5
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
yarl==1.8.2
